requirements.txt
faiss-cpu
pickle5
langchain-community
transformers
sentence-transformers
torch
pymupdf